+++ internal/daemon/service_other.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 // Copyright (C) 2026 Hassan Khurram
 
-//go:build !linux && !darwin
+//go:build !linux && !darwin && !windows
 
 package daemon